#80 #86: deleted redundent files
